updating reader.py
