+++ holoclean/learning/softmax.py
@@ -17,7 +17,6 @@ class LogReg(torch.nn.Module):
         """
         Initializes weight tensor with random values
         ties init and dc weights if specified
-
         :return: Null
         """
         torch.manual_seed(42)
@@ -49,11 +48,10 @@ class LogReg(torch.nn.Module):
             self.weight_tensors.append(signals_W)
         return
 
-    def __init__(self, featurizers, input_dim_non_dc, input_dim_dc, output_dim,
+    def __init__(self, featurizers, output_dim,
                  tie_init, tie_dc):
         """
         Constructor for our logistic regression
-
         :param input_dim_non_dc: number of init + cooccur features
         :param input_dim_dc: number of dc features
         :param output_dim: number of classes
@@ -64,8 +62,6 @@ class LogReg(torch.nn.Module):
 
         self.featurizers = featurizers
 
-        self.input_dim_non_dc = input_dim_non_dc
-        self.input_dim_dc = input_dim_dc
         self.output_dim = output_dim
 
         self.tie_init = tie_init
@@ -76,7 +72,6 @@ class LogReg(torch.nn.Module):
     def forward(self, X, index, mask):
         """
         Runs the forward pass of our logreg.
-
         :param X: values of the features
         :param index: indices to mask at
         :param mask: tensor to remove possibility of choosing unused class
@@ -120,20 +115,14 @@ class SoftMax:
     def __init__(self, session, X_training):
         """
         Constructor for our softmax model
-
         :param X_training: x tensor used for training the model
         :param session: session object
-
         """
         self.session = session
         self.dataengine = session.holo_env.dataengine
         self.dataset = session.dataset
         self.holo_obj = session.holo_env
         self.spark_session = self.holo_obj.spark_session
-        query = "SELECT COUNT(*) AS dc FROM " + \
-                self.dataset.table_specific_name("Feature_id_map") + \
-                " WHERE Type = 'DC'"
-        self.DC_count = self.dataengine.query(query, 1).collect()[0].dc
         dataframe_offset = self .dataengine.get_table_to_dataframe(
             "Dimensions_clean", self.dataset)
         list = dataframe_offset.collect()
@@ -164,7 +153,6 @@ class SoftMax:
     def _setupY(self):
         """
         Initializes a y tensor to compare to our model's output
-
         :return: Null
         """
         possible_values = self.dataengine .get_table_to_dataframe(
@@ -179,9 +167,7 @@ class SoftMax:
     def _setupX(self, sparse=0):
         """
         Initializes an X tensor of features for prediction
-
         :param sparse: 0 if dense tensor, 1 if sparse
-
         :return: Null
         """
         feature_table = self .dataengine.get_table_to_dataframe(
@@ -209,9 +195,7 @@ class SoftMax:
     def setuptrainingX(self, sparse=0):
         """
         Initializes an X tensor of features for training
-
         :param sparse: 0 if dense tensor, 1 if sparse
-
         :return: x tensor of features
         """
         dataframe_offset = self.dataengine.get_table_to_dataframe(
@@ -252,11 +236,9 @@ class SoftMax:
     def setupMask(self, clean=1, N=1, L=1):
         """
         Initializes a masking tensor for ignoring impossible classes
-
         :param clean: 1 if clean cells, 0 if don't-know
         :param N: number of examples
         :param L: number of classes
-
         :return: masking tensor
         """
         lookup = "Kij_lookup_clean" if clean else "Kij_lookup_dk"
@@ -274,40 +256,34 @@ class SoftMax:
             self.testmask = mask
         return mask
 
-    def build_model(self,  featurizers, input_dim_non_dc, input_dim_dc,
+    def build_model(self,  featurizers,
                     output_dim, tie_init=True, tie_DC=True):
         """
         Initializes the logreg part of our model
-
         :param input_dim_non_dc: number of init + cooccur features
         :param featurizers: list of featurizers
         :param input_dim_dc: number of dc features
         :param output_dim: number of classes
         :param tie_init: boolean to decide weight tying for init features
         :param tie_DC: boolean to decide weight tying for dc features
-
         :return: newly created LogReg model
         """
         model = LogReg(
             featurizers,
-            input_dim_non_dc,
-            input_dim_dc,
             output_dim,
             tie_init,
-            tie_DC,)
+            tie_DC)
         return model
 
     def train(self, model, loss, optimizer, x_val, y_val, mask=None):
         """
         Trains our model on the clean cells
-
         :param model: logistic regression model
         :param loss: loss function used for evaluating performance
         :param optimizer: optimizer for our neural net
         :param x_val: x tensor - features
         :param y_val: y tensor - output for comparison
         :param mask: masking tensor
-
         :return: cost of traininng
         """
         x = Variable(x_val, requires_grad=False)
@@ -338,11 +314,9 @@ class SoftMax:
     def predict(self, model, x_val, mask=None):
         """
         Runs our model on the test set
-
         :param model: trained logreg model
         :param x_val: test x tensor
         :param mask: masking tensor to restrict domain
-
         :return: predicted classes with probabilities
         """
         x = Variable(x_val, requires_grad=False)
@@ -360,12 +334,11 @@ class SoftMax:
     def logreg(self, featurizers):
         """
         Trains our model on clean cells and predicts vals for clean cells
-
         :return: predictions
         """
         n_examples, n_features, n_classes = self.X.size()
         self.model = self.build_model(
-            featurizers, self.M - self.DC_count, self.DC_count, n_classes)
+            featurizers, n_classes)
         loss = torch.nn.CrossEntropyLoss(size_average=True)
         optimizer = optim.SGD(
             self.model.parameters(),
@@ -398,9 +371,7 @@ class SoftMax:
     def save_prediction(self, Y):
         """
         Stores our predicted values in the database
-
         :param Y: tensor with probability for each class
-
         :return: Null
         """
         k_inferred = self.session.holo_env.k_inferred
@@ -447,7 +418,6 @@ class SoftMax:
     def log_weights(self):
         """
         Writes weights in the logger
-
         :return: Null
         """
         self.model.concat_weights()
